Added logic if the user doesn't already have acc
Previously it wouldn't retain subscriptions set to the user if he didn't
already have an account.  The script now checks to see if the user id is
zero (which means that the operation is creating a new user).

Next it checks for the next auto increment value in the database and
assigns that number to the customer.

##Only possible issue would be if two admins were to do this at the
exact same second.  I don't forsee this happening if at all.